♻️ FindBarsByRequestList 모듈화
findBarsByRequestList 모듈화
