+++ src/main/java/com/Ohsul/Ohsul/service/NearAlcoholService.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.*;
 import org.springframework.stereotype.*;
 
 import java.util.*;
-import java.util.stream.*;
 
 @Service
 public class NearAlcoholService {
@@ -19,6 +18,8 @@ public class NearAlcoholService {
 
   @Autowired
   BarEntityToDTOConverter barEntityToDTOConverter;
+  @Autowired
+  FindBarByRequestList findBarByRequestList;
 
   public NearAlcoholService(BarRepository barRepository, ReviewRepository reviewRepository) {
 
@@ -27,33 +28,6 @@ public class NearAlcoholService {
   }
 
   public List<BarListDTO> findBarsByRequestList(List<BarSearchDTO> requests) {
-    Set<BarEntity> barSet = new HashSet<>();
-    for (BarSearchDTO request : requests) {
-      BarEntity bar = null;
-
-      if (request.getTelephone() != null && !request.getTelephone().isBlank()) {
-        bar = barRepository.findByTelephone(request.getTelephone());
-      }
-
-      if (bar == null && request.getBarName() != null && !request.getBarName().isBlank()) {
-        bar = barRepository.findByBarName(request.getBarName());
-      }
-
-      if (bar == null) {
-        bar = new BarEntity();
-        bar.setTelephone(request.getTelephone());
-        bar.setBarName(request.getBarName());
-        bar.setRoadAddress(request.getRoadAddress());
-        barRepository.save(bar);
-      }
-
-      barSet.add(bar);
-    }
-    return barSet.stream().map(this::convertEntityToDto).collect(Collectors.toList());
-  }
-
-  private BarListDTO convertEntityToDto(BarEntity barentity) {
-    return barEntityToDTOConverter.convertEntityToDto(barentity);
-
+    return findBarByRequestList.findBarsByRequestList(requests);
   }
 }
